# backend/ent/coolroomtype.go
// Code generated by entc, DO NOT EDIT.

package ent

import (
	"fmt"
	"strings"

	"github.com/B6113759/app/ent/coolroomtype"
	"github.com/facebookincubator/ent/dialect/sql"
)

// CoolroomType is the model entity for the CoolroomType schema.
type CoolroomType struct {
	config `json:"-"`
	// ID of the ent.
	ID int `json:"id,omitempty"`
	// CoolroomtypeName holds the value of the "coolroomtype_name" field.
	CoolroomtypeName string `json:"coolroomtype_name,omitempty"`
	// Edges holds the relations/edges for other nodes in the graph.
	// The values are being populated by the CoolroomTypeQuery when eager-loading is set.
	Edges CoolroomTypeEdges `json:"edges"`
}

// CoolroomTypeEdges holds the relations/edges for other nodes in the graph.
type CoolroomTypeEdges struct {
	// Coolrooms holds the value of the coolrooms edge.
	Coolrooms []*Coolroom
	// loadedTypes holds the information for reporting if a
	// type was loaded (or requested) in eager-loading or not.
	loadedTypes [1]bool
}

// CoolroomsOrErr returns the Coolrooms value or an error if the edge
// was not loaded in eager-loading.
func (e CoolroomTypeEdges) CoolroomsOrErr() ([]*Coolroom, error) {
	if e.loadedTypes[0] {
		return e.Coolrooms, nil
	}
	return nil, &NotLoadedError{edge: "coolrooms"}
}

// scanValues returns the types for scanning values from sql.Rows.
func (*CoolroomType) scanValues() []interface{} {
	return []interface{}{
		&sql.NullInt64{},  // id
		&sql.NullString{}, // coolroomtype_name
	}
}

// assignValues assigns the values that were returned from sql.Rows (after scanning)
// to the CoolroomType fields.
func (ct *CoolroomType) assignValues(values ...interface{}) error {
	if m, n := len(values), len(coolroomtype.Columns); m < n {
		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
	}
	value, ok := values[0].(*sql.NullInt64)
	if !ok {
		return fmt.Errorf("unexpected type %T for field id", value)
	}
	ct.ID = int(value.Int64)
	values = values[1:]
	if value, ok := values[0].(*sql.NullString); !ok {
		return fmt.Errorf("unexpected type %T for field coolroomtype_name", values[0])
	} else if value.Valid {
		ct.CoolroomtypeName = value.String
	}
	return nil
}

// QueryCoolrooms queries the coolrooms edge of the CoolroomType.
func (ct *CoolroomType) QueryCoolrooms() *CoolroomQuery {
	return (&CoolroomTypeClient{config: ct.config}).QueryCoolrooms(ct)
}

// Update returns a builder for updating this CoolroomType.
// Note that, you need to call CoolroomType.Unwrap() before calling this method, if this CoolroomType
// was returned from a transaction, and the transaction was committed or rolled back.
func (ct *CoolroomType) Update() *CoolroomTypeUpdateOne {
	return (&CoolroomTypeClient{config: ct.config}).UpdateOne(ct)
}

// Unwrap unwraps the entity that was returned from a transaction after it was closed,
// so that all next queries will be executed through the driver which created the transaction.
func (ct *CoolroomType) Unwrap() *CoolroomType {
	tx, ok := ct.config.driver.(*txDriver)
	if !ok {
		panic("ent: CoolroomType is not a transactional entity")
	}
	ct.config.driver = tx.drv
	return ct
}

// String implements the fmt.Stringer.
func (ct *CoolroomType) String() string {
	var builder strings.Builder
	builder.WriteString("CoolroomType(")
	builder.WriteString(fmt.Sprintf("id=%v", ct.ID))
	builder.WriteString(", coolroomtype_name=")
	builder.WriteString(ct.CoolroomtypeName)
	builder.WriteByte(')')
	return builder.String()
}

// CoolroomTypes is a parsable slice of CoolroomType.
type CoolroomTypes []*CoolroomType

func (ct CoolroomTypes) config(cfg config) {
	for _i := range ct {
		ct[_i].config = cfg
	}
}